# sync-server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fractonica API
  version: current
  summary: Server-authoritative journal, event, and synchronization API.
  description: |
    The Fractonica API is the canonical server contract.
    It is multi-user, cursor-synchronized, and intended for the first-party iOS
    client as well as user-authorized automation agents.

    Authentication uses either a short-lived JWT access token (`JwtBearer`) or
    a user-issued API key (`ApiKeyBearer`) in the `Authorization: Bearer ...`
    header. API keys are bound to one user and one logical device. Operations
    carrying `x-required-scopes` enforce those scopes for API-key callers; JWT
    sessions act with the authenticated user's full authority.

    Production deployments must expose the API only over HTTPS/TLS and reject
    plain HTTP at the edge. Passwords and bearer credentials are otherwise
    observable in transit. Plain HTTP is for loopback development only.

    Records are public by default. Public record payloads are readable and may
    be rendered from a template. Private record payloads and private
    media are encrypted by the client before upload; the server stores only an
    authenticated ciphertext envelope and intentionally cannot render, index,
    or embed their plaintext.

    All timestamps are RFC 3339 instants. Record API identifiers are globally
    unique five-character Base64URL strings; their owner-only `originId` UUIDs
...
    are UUIDs.
...

...
    immediately retrying.

...
  license:
...
servers:
...
security:
  - JwtBearer: []
  - ApiKeyBearer: []
...
paths:
...
  /public/users/{login}:
    get:
      operationId: getPublicUser
      tags: [Social]
      summary: Resolve a public profile by stable login
      description: |
        Returns 404 while the user is disabled. A `user` lifecycle item in the
        public activity stream uses this URL: `delete` removes the public
        projection and `upsert` makes it fetchable again.
      security: []
      parameters:
        - name: login
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/LoginName'
      responses:
        '200':
          description: Public user profile and public-resource counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserProfile'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [Social]
      summary: List the authenticated user's subscriptions
      x-required-scopes: [subscriptions:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Private owner subscription page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /subscriptions/{targetUserId}:
    parameters:
      - name: targetUserId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      operationId: putSubscription
      tags: [Social]
      summary: Subscribe to a public user or update feed kinds
      x-required-scopes: [subscriptions:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionInput'
      responses:
        '200':
          description: Existing subscription updated or retained.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '201':
          description: Subscription created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: deleteSubscription
      tags: [Social]
      summary: Stop following a public user
      x-required-scopes: [subscriptions:write]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Subscription removed and owner tombstone emitted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /public/activity:
    get:
      operationId: listPublicActivity
      tags: [Public activity]
      summary: Read the global public activity cursor
      description: |
        Returns append-only user lifecycle controls and record/event revision
        notifications in commit-safe sequence order. Payloads are fetched from
        each `resourceUrl`; activity items never embed resource payloads.

        For `resourceType: user`, `delete` means the profile is no longer
        public and clients must hide or purge every cached public resource from
        that actor. `upsert` means the profile is public again and clients must
        refetch the profile plus the actor-filtered public record/event lists.
        User lifecycle controls are automatically included when records or
        events are selected so a filtered cursor cannot skip an invalidation.

        User lifecycle is status-based and this API does not hard-delete
        users. Administrative hard deletion is a destructive feed-reset
        operation: consumers must discard cursors and rebuild afterward.
      security: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PublicActivitySnapshot'
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PublicActivityResourceTypeFilter'
      responses:
        '200':
          description: Global public activity page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicActivityPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /activity:
    get:
      operationId: listFollowingActivity
      tags: [Public activity]
      summary: Read public activity from subscribed users
      x-required-scopes: [subscriptions:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PublicActivitySnapshot'
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PublicActivityResourceTypeFilter'
      responses:
        '200':
          description: Subscription-filtered public activity page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicActivityPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    JwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived Ed25519-signed access JWT issued by login or refresh. The
        token includes `iss`, `aud`, `sub`, `exp`, and `kid` claims.
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: Fractonica API key
      description: |
        User-issued secret beginning with `exk_`. Keys are stored hashed,
        device-bound, scoped, independently expirable, and revocable. The
        plaintext secret appears only in the initial creation response; a replay
        returns the created `apiKeyId`, never the secret.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the immediately preceding page.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-generated identifier for one logical mutation. It must be reused
        for retries of the same body and must not be reused for another body.
        Most operations replay their original response. API-key creation is the
        deliberate exception: replay returns a 409 problem containing the
        created `apiKeyId` because its plaintext secret is returned only once.
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: '^[\x21-\x7E]+$'
      example: 01J2V7YCB9Q3WZ2P6AQ03M6DN8
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag of the revision being mutated.
      schema:
        type: string
        minLength: 3
        maxLength: 200
      example: '"record-aB_9Z-r4"'
...
    PublicActivityResourceTypeFilter:
      name: resourceType
      in: query
      description: |
        Repeat to select users, records, and/or events. Selecting records or
        events automatically includes user lifecycle controls required to
        invalidate cached resources safely.
      style: form
      explode: true
      schema:
        type: array
        maxItems: 3
        uniqueItems: true
        items:
          $ref: '#/components/schemas/PublicActivityResourceType'
    PublicActivitySnapshot:
      name: snapshot
      in: query
      description: |
        Set to `latest` only when `cursor` is omitted to receive at most the
        newest `limit` matching notifications, returned in canonical ascending
        sequence order. The page intentionally omits older history, reports
        `hasMore: false`, and returns a normal cursor anchored at the
        repeatable-read high-water mark. Resume live changes by sending that
        cursor on a subsequent request without `snapshot`.
      schema:
        type: string
        enum: [latest]
  headers:
    ETag:
      description: Strong opaque validator for the returned resource revision.
      schema:
        type: string
      example: '"record-aB_9Z-r4"'
    RequestId:
      description: Server-assigned request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: The request syntax, parameter combination, or cursor is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.exeligmos.app/problems/invalid-request
            title: Invalid request
            status: 400
            detail: The occurredAfter timestamp must precede occurredBefore.
            code: invalid_request
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    Unauthorized:
      description: Authentication is missing, expired, revoked, or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="exeligmos"
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the scope or ownership required for this operation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The resource is absent or is not visible to this caller.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The mutation conflicts with current state or idempotency history.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    PreconditionFailed:
      description: "`If-Match` does not match the current resource revision."
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    UnprocessableContent:
      description: The request is syntactically valid but violates domain constraints.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    ServiceUnavailable:
      description: A database statement or lock exceeded its safety deadline; the operation may be retried.
      headers:
        Retry-After:
          description: Minimum number of seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
          example: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:exeligmos:problem:database-timeout
            title: Service Unavailable
            status: 503
            detail: The database could not complete the request within its safety deadline.
            code: database_timeout
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    TooManyRequests:
      description: Rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details with stable application extensions.
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type URI.
        title:
          type: string
          minLength: 1
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
      additionalProperties: true
    FieldError:
      type: object
      required: [path, code, message]
      properties:
        path:
          type: string
          description: JSON Pointer to the invalid value.
          maxLength: 500
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        message:
          type: string
          minLength: 1
          maxLength: 1000
      additionalProperties: false
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              minItems: 1
              maxItems: 100
              items:
                $ref: '#/components/schemas/FieldError'
...
    CursorPageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: string
          minLength: 1
          maxLength: 2048
        hasMore:
          type: boolean
...
    LoginName:
      type: string
      description: Case-insensitive account login stored in normalized form.
      minLength: 3
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
...
    SarosAnchor:
      type: integer
      minimum: 1
      maximum: 180
      default: 141
      description: Saros series used as the user's temporal pulse anchor.
    PublicUserSummary:
      type: object
      required: [id, login, displayName, sarosAnchor]
      properties:
        id:
          type: string
          format: uuid
        login:
          $ref: '#/components/schemas/LoginName'
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        sarosAnchor:
          $ref: '#/components/schemas/SarosAnchor'
      additionalProperties: false
    PublicUserProfile:
      allOf:
        - $ref: '#/components/schemas/PublicUserSummary'
        - type: object
          required: [createdAt, publicRecordCount, publicEventCount, followerCount]
          properties:
            createdAt:
              type: string
              format: date-time
            publicRecordCount:
              type: integer
              minimum: 0
            publicEventCount:
              type: integer
              minimum: 0
            followerCount:
              type: integer
              minimum: 0
      unevaluatedProperties: false
...
    SubscriptionInput:
      type: object
      properties:
        includeRecords:
          type: boolean
          default: true
        includeEvents:
          type: boolean
          default: true
      additionalProperties: false
    Subscription:
      type: object
      required:
        [id, userId, targetUser, includeRecords, includeEvents, revision, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        targetUser:
          $ref: '#/components/schemas/SubscriptionTargetUserSummary'
        includeRecords:
          type: boolean
        includeEvents:
          type: boolean
        revision:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    SubscriptionTargetUserSummary:
      type: object
      required: [id, login, displayName, sarosAnchor, status]
      properties:
        id:
          type: string
          format: uuid
        login:
          $ref: '#/components/schemas/LoginName'
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        sarosAnchor:
          $ref: '#/components/schemas/SarosAnchor'
        status:
          type: string
          enum: [active, disabled]
          description: Current account availability for this existing subscription target.
      additionalProperties: false
    SubscriptionPage:
      allOf:
        - $ref: '#/components/schemas/CursorPageMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Subscription'
...
    RecordPublicId:
      type: string
      minLength: 5
      maxLength: 5
      pattern: '^[A-Za-z0-9_-]{5}$'
      description: Globally unique, case-sensitive Base64URL record identifier.
    ResourceIdentifier:
      description: >-
        Record resources use RecordPublicId; all other resource types use UUIDs.
      oneOf:
        - $ref: '#/components/schemas/RecordPublicId'
        - type: string
          format: uuid
...
    PublicActivityResourceType:
      type: string
      enum: [user, record, event]
    PublicActivityItem:
      type: object
      description: |
        Identifier-only public activity notification. A user `delete` removes
        the actor's public projection and all cached public resources; a user
        `upsert` restores the projection and requires actor-filtered refetches.
        The user resource ID is the actor ID, while its resource URL uses the
        stable public login.
      required:
        [sequence, publishedAt, actor, resourceType, resourceId, operation, revision, resourceUrl]
      properties:
        sequence:
          type: integer
          format: int64
          minimum: 1
        publishedAt:
          type: string
          format: date-time
        actor:
          $ref: '#/components/schemas/PublicUserSummary'
        resourceType:
          $ref: '#/components/schemas/PublicActivityResourceType'
        resourceId:
          $ref: '#/components/schemas/ResourceIdentifier'
        operation:
          type: string
          enum: [upsert, delete]
        revision:
          type: integer
          format: int64
          minimum: 1
        resourceUrl:
          type: string
          format: uri-reference
      additionalProperties: false
    PublicActivityPage:
      type: object
      required: [data, nextCursor, hasMore]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PublicActivityItem'
        nextCursor:
          type: string
          minLength: 1
          maxLength: 2048
        hasMore:
          type: boolean
      additionalProperties: false
...